application/pyproject.toml
[tool.poetry]
name = "crypto-hash-server"
version = "0.1.0"
description = "An application for hash/key generation"
authors = ["Oleksii-Op <oleksii.oprish@gmail.com>"]
readme = "README.md"
license = "GNU GPL v2"
packages = [{include = "crypto_hash_server"}]

[tool.poetry.dependencies]
python = "3.12.7"
cryptography = "^44.0.2"
fastapi = {extras = ["standard"], version = "^0.115.11"}
pydantic-settings = "^2.8.1"
pwdlib = {extras = ["argon2", "bcrypt"], version = "^0.2.1"}
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation-fastapi = "^0.51b0"
opentelemetry-exporter-otlp = "^1.30.0"
prometheus-fastapi-instrumentator = "^7.0.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
mkdocs-material = "^9.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini.options]
filterwarnign = ["ignore::DeprecationWarning"]

[tool.coverage.run]
omit = [
    "core/config.py",
    "core/schemas.py",
    "response_docs/*",
    "tests/*",
    "main.py",
    "__init__.py"
]